+++ package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^6.1.3",
+    "@mui/joy": "^5.0.0-beta.48",
     "@mui/material": "^6.1.3",
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^13.5.0",
