+++ src/components/modal/modal.js
@@ -1,10 +1,14 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
 import styles from './modal.module.css';
 import ModalOverlay from '../modal/modal-overlay';
 import ClearIcon from '@mui/icons-material/Clear';
 const modalRoot = document.getElementById('modal-root');
 const Modal = ({ children, onClose }) => {
+  // const [open, setOpen] = useState(false);
+
+  // const handleClose = () => setOpen(false);
+  // const handleOpen = () => setOpen(true);
   useEffect(() => {
     const handleClose = (event) => {
       if (event.key === 'Escape') {
@@ -23,9 +27,7 @@ const Modal = ({ children, onClose }) => {
       <ModalOverlay onClose={onClose} />
       <div className={styles.modal}>
         <div onClick={onClose}>
-          <p>
-            <ClearIcon className={styles.clearIcon} />
-          </p>
+          <ClearIcon className={styles.clearIcon} />
         </div>
         {children}
       </div>
